Securizada aplicación por parte de pistas reviews y juegousuario para que solamente los propietarios puedan borrar sus entradas

+++ Player2REST/src/main/java/es/iespuertodelacruz/daniel/Player2REST/controller/v1/JuegoUsuarioRESTv1.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -16,9 +17,11 @@ import es.iespuertodelacruz.daniel.Player2REST.dto.JuegoUsuarioDTO;
 import es.iespuertodelacruz.daniel.Player2REST.entity.JuegoUsuario;
 import es.iespuertodelacruz.daniel.Player2REST.entity.Usuario;
 import es.iespuertodelacruz.daniel.Player2REST.entity.Videojuego;
+import es.iespuertodelacruz.daniel.Player2REST.security.GestorDeJWT;
 import es.iespuertodelacruz.daniel.Player2REST.service.JuegoUsuarioService;
 import es.iespuertodelacruz.daniel.Player2REST.service.UsuarioService;
 import es.iespuertodelacruz.daniel.Player2REST.service.VideojuegoService;
+import io.jsonwebtoken.Claims;
 
 @RestController
 @RequestMapping("/api/v1/juegousuario")
@@ -33,11 +36,19 @@ public class JuegoUsuarioRESTv1 {
 	
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> delete(@PathVariable Integer id) {
+	public ResponseEntity<?> delete(@PathVariable Integer id, @RequestHeader (name="Authorization") String token) {
 		Optional<JuegoUsuario> optM = juegousuarioService.findById(id);
 		if (optM.isPresent()) {
-			juegousuarioService.deleteById(id);
-			return ResponseEntity.ok("juegousuario borrado");
+			GestorDeJWT gestorDeJwt = GestorDeJWT.getInstance();
+			String tokenCorregido = token.split(" ")[1].trim();
+			Claims claimsGestor = gestorDeJwt.getClaims(tokenCorregido);
+			String username = claimsGestor.getSubject();
+			if (optM.get().getUsuario().getNombre().equals(username)) {
+				juegousuarioService.deleteById(id);
+				return ResponseEntity.ok("juegousuario borrado");
+			} else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Solo el usuario propietario puede borrar sus entradas");
+			}
 		} else {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("el id del juegousuario no existe");
 		}
@@ -59,32 +70,38 @@ public class JuegoUsuarioRESTv1 {
 	}
 */
 	@PostMapping
-	public ResponseEntity<?> saveJuegoUsuario(@RequestBody JuegoUsuarioDTO juegousuarioDto) {
+	public ResponseEntity<?> saveJuegoUsuario(@RequestBody JuegoUsuarioDTO juegousuarioDto, @RequestHeader (name="Authorization") String token) {
 		
-		JuegoUsuario juegousuario = new JuegoUsuario();
-		Optional<Videojuego> videojuego = videojuegoService.findById(juegousuarioDto.getVideojuego().getId());
-		Optional<Usuario> usuario = usuarioService.findById(juegousuarioDto.getUsuario().getId());
-		if (videojuego.get() != null && usuario.get() != null) {
-			juegousuario.setCompletado((byte) (juegousuarioDto.isCompletado() ? 1 : 0));
-			juegousuario.setHoras(juegousuarioDto.getHoras());
-			juegousuario.setVideojuego(videojuego.get());
-			juegousuario.setPuntuacion(juegousuarioDto.getPuntuacion());
-			juegousuario.setUsuario(usuario.get());
-			JuegoUsuario juegousuarioC = null;
-			try {
-				juegousuarioC = juegousuarioService.save(juegousuario);
-			} catch (Exception e) {
-				e.printStackTrace();
+			JuegoUsuario juegousuario = new JuegoUsuario();
+			Optional<Videojuego> videojuego = videojuegoService.findById(juegousuarioDto.getVideojuego().getId());
+			Optional<Usuario> usuario = usuarioService.findById(juegousuarioDto.getUsuario().getId());
+			if (videojuego.get() != null && usuario.get() != null) {
+				GestorDeJWT gestorDeJwt = GestorDeJWT.getInstance();
+				String tokenCorregido = token.split(" ")[1].trim();
+				Claims claimsGestor = gestorDeJwt.getClaims(tokenCorregido);
+				String username = claimsGestor.getSubject();
+				if (usuario.get().getNombre().equals(username)) {
+					juegousuario.setCompletado((byte) (juegousuarioDto.isCompletado() ? 1 : 0));
+					juegousuario.setHoras(juegousuarioDto.getHoras());
+					juegousuario.setVideojuego(videojuego.get());
+					juegousuario.setPuntuacion(juegousuarioDto.getPuntuacion());
+					juegousuario.setUsuario(usuario.get());
+					JuegoUsuario juegousuarioC = null;
+					try {
+						juegousuarioC = juegousuarioService.save(juegousuario);
+					} catch (Exception e) {
+						e.printStackTrace();
+					}
+					if (juegousuarioC != null) {
+						return new ResponseEntity<>(juegousuarioC, HttpStatus.OK);
+					} else {
+						return new ResponseEntity<>("Ya existe esa combinación de valores (Nick, Password)", HttpStatus.CONFLICT);
+					} 
+				} else{
+					return new ResponseEntity<>("Solo el usuario propietario puede crear nuevas entradas", HttpStatus.UNAUTHORIZED);
+				}
+			} else{
+				return new ResponseEntity<>("No se encuentra el videojuego o el usuario en la bbdd", HttpStatus.CONFLICT);
 			}
-			if (juegousuarioC != null) {
-				return new ResponseEntity<>(juegousuarioC, HttpStatus.OK);
-			} else {
-				return new ResponseEntity<>("Ya existe esa combinación de valores (Nick, Password)", HttpStatus.CONFLICT);
-			}
-		} else{
-			return new ResponseEntity<>("No se encuentra el videojuego o el usuario en la bbdd", HttpStatus.CONFLICT);
-		}
-		
-
 	}
 }
