+++ Player2REST/src/main/java/es/iespuertodelacruz/daniel/Player2REST/controller/v1/PistaRESTv1.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -18,9 +19,11 @@ import es.iespuertodelacruz.daniel.Player2REST.dto.PistaDTO;
 import es.iespuertodelacruz.daniel.Player2REST.entity.Pista;
 import es.iespuertodelacruz.daniel.Player2REST.entity.Usuario;
 import es.iespuertodelacruz.daniel.Player2REST.entity.Videojuego;
+import es.iespuertodelacruz.daniel.Player2REST.security.GestorDeJWT;
 import es.iespuertodelacruz.daniel.Player2REST.service.PistaService;
 import es.iespuertodelacruz.daniel.Player2REST.service.UsuarioService;
 import es.iespuertodelacruz.daniel.Player2REST.service.VideojuegoService;
+import io.jsonwebtoken.Claims;
 
 @RestController
 @RequestMapping("/api/v1/pista")
@@ -34,11 +37,19 @@ public class PistaRESTv1 {
 	UsuarioService usuarioService;
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> delete(@PathVariable Integer id) {
+	public ResponseEntity<?> delete(@PathVariable Integer id, @RequestHeader (name="Authorization") String token) {
 		Optional<Pista> optM = pistaService.findById(id);
 		if (optM.isPresent()) {
-			pistaService.deleteById(id);
-			return ResponseEntity.ok("pista borrado");
+			GestorDeJWT gestorDeJwt = GestorDeJWT.getInstance();
+			String tokenCorregido = token.split(" ")[1].trim();
+			Claims claimsGestor = gestorDeJwt.getClaims(tokenCorregido);
+			String username = claimsGestor.getSubject();
+			if (optM.get().getUsuario().getNombre().equals(username)) {
+				pistaService.deleteById(id);
+				return ResponseEntity.ok("pista borrado");
+			} else {
+				return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Solo el usuario propietario puede borrar sus entradas");
+			}
 		} else {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body("el id del pista no existe");
 		}
@@ -61,29 +72,38 @@ public class PistaRESTv1 {
 	}
 */
 	@PostMapping
-	public ResponseEntity<?> savePista(@RequestBody PistaDTO pistaDto) {
+	public ResponseEntity<?> savePista(@RequestBody PistaDTO pistaDto, @RequestHeader (name="Authorization") String token) {
 		
 		Pista pista = new Pista();
 		Optional<Videojuego> videojuego = videojuegoService.findById(pistaDto.getVideojuego().getId());
 		Optional<Usuario> usuario = usuarioService.findById(pistaDto.getUsuario().getId());
 		if (videojuego.get() != null && usuario.get() != null) {
-			pista.setContenido(pistaDto.getContenido());
-			pista.setFecha(BigInteger.valueOf(new Date().getTime()));
-			pista.setUsuario(usuario.get());
-			pista.setVideojuego(videojuego.get());
-			pista.setTitulo(pistaDto.getTitulo());
-			Pista pistaC = null;
-			try {
-				pistaC = pistaService.save(pista);
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-			if (pistaC != null) {
-				return new ResponseEntity<>(pistaC, HttpStatus.OK);
+			GestorDeJWT gestorDeJwt = GestorDeJWT.getInstance();
+			String tokenCorregido = token.split(" ")[1].trim();
+			Claims claimsGestor = gestorDeJwt.getClaims(tokenCorregido);
+			String username = claimsGestor.getSubject();
+			if (usuario.get().getNombre().equals(username)) {
+				pista.setContenido(pistaDto.getContenido());
+				pista.setFecha(BigInteger.valueOf(new Date().getTime()));
+				pista.setUsuario(usuario.get());
+				pista.setVideojuego(videojuego.get());
+				pista.setTitulo(pistaDto.getTitulo());
+				Pista pistaC = null;
+				try {
+					pistaC = pistaService.save(pista);
+				} catch (Exception e) {
+					e.printStackTrace();
+				}
+				if (pistaC != null) {
+					return new ResponseEntity<>(pistaC, HttpStatus.OK);
+				} else {
+					return new ResponseEntity<>("Ya existe esa combinación de valores (Nick, Password)", HttpStatus.CONFLICT);
+				}
 			} else {
-				return new ResponseEntity<>("Ya existe esa combinación de valores (Nick, Password)", HttpStatus.CONFLICT);
+				return new ResponseEntity<>("Solo el usuario propietario puede crear nuevas entradas", HttpStatus.UNAUTHORIZED);
 			}
-		} else{
+		} 
+		else{
 			return new ResponseEntity<>("No se encuentra el videojuego o el usuario en la bbdd", HttpStatus.CONFLICT);
 		}
